SomeChanges

+++ WpfApp2/Model/DTO/PropertyDTO.cs
@@ -27,7 +27,7 @@
         public DateTime OrderDate { get; set; }
         public int Status { get; set; }
         public string StatusName { get; set; }
-        public double Quantity { get; set; }
-        public double Price { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal Price { get; set; }
     }
 }
