+++ WpfApp2/ViewModel/Model/Property.cs
@@ -27,8 +27,8 @@
         public DateTime OrderDate { get; set; }
         public int Status { get; set; }
         public string StatusName { get; set; }
-        public double Quantity { get; set; }
-        public double Price { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal Price { get; set; }
         public Property(int factoryNumber, string? name,
             int inventoryNumber, int invoiceId,  
             int bookId, int orderBookId,string? formId,
